refactor: renamed f56_module to float_56
