+++ vm/src/float_56.rs
